Eliminate infinite recursive loop in EMF output.
ByteCountOutputStream subclasses ByteOrderOutputStream, which is a descendant
of CompressableOutputStream. CompressableOutputStream.write(int b) calls
an overridden writeSingleByte(). When the object instance is an
EMFOutputStream, there's an infinite recursive loop:
  EMFOutputStream.write(int b) (inherited from CompressbleOutputStream)
    calls writeSingleByte(int b) (resolves to ByteCountOutputStream's method)
      calls super.write(int b)  (back up to CompressibleOutputStream)
And so on, forever. Changing super.write(b) to super.writeSingleByte(b) fixes
the problem, as CompressableOutputStream implements a protected
writeSingleByte.